// frontend/src/generated/model/getSearchSource.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * AnthoLume API v1
 * REST API for AnthoLume document management system
 * OpenAPI spec version: 1.0.0
 */

export type GetSearchSource = typeof GetSearchSource[keyof typeof GetSearchSource];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetSearchSource = {
  LibGen: 'LibGen',
  Annas_Archive: 'Annas Archive',
} as const;